fix:change to indonesian lang in FE and confirm password in store is not required
