+++ frontend/src/app/master-data/partner/tambah/page.tsx
