+++ frontend/src/app/pelanggan/bank/tambah/page.tsx
