+++ frontend/src/app/pengguna/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import React, { use, useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
+import React, { useState, useEffect } from "react";
+import { useRouter, useParams } from "next/navigation";
 import api from "@/services/api";
 import { cn } from "@/lib/utils";
 import { ProtectedRoute } from "@/components/ProtectedRoute";
@@ -14,6 +14,7 @@ import {
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
+import { User } from "@/types/user";
 import {
   Select,
   SelectContent,
@@ -23,26 +24,43 @@ import {
 } from "@/components/ui/select";
 
 // ------------------- Component -------------------
-export default function CreateRole() {
+export default function CreateUser() {
   const router = useRouter();
+  const params = useParams();
+  const { id } = params;
   const [saving, setSaving] = useState(false);
-  const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(false)
+  const [error, setError] = useState<Error | null>(null);
   const [roleResponse, setRoleResponse] = useState<any[]>([]);
 
-  const [formData, setFormData] = useState<{ name: string, username: string, email: string, password: string, role_id: string }>({
+
+  const [formData, setFormData] = useState<{ name: string, username: string, email: string, password: string, password_confirmation: string, current_password: string, role_id: string }>({
     name: "",
     username: "",
     email: "",
     password: "",
+    password_confirmation: "",
+    current_password: "",
     role_id: "",
   });
 
-  const handleInputChange = (field: keyof typeof formData, value: string) => {
-    setFormData((prev) => ({ ...prev, [field]: value }));
-  };
-
   useEffect(() => {
-    const fetchRole = async () => {
+    const fetchUser = async () => {
+      try {
+        setLoading(true);
+        const response = await api.get(`/users/${id}`);
+        const user = response.data.user;
+        setFormData({...formData, name: user.name, username: user.username, email: user.email, role_id: user.role_id.toString()});
+        console.log(user);
+
+      } catch (err: any) {
+        setError(err);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+     const fetchRole = async () => {
       try {
         setLoading(true);
         const roleResponse = await api.get("/roles");
@@ -55,7 +73,13 @@ export default function CreateRole() {
     };
 
     fetchRole();
-  }, []);
+
+    if (id) fetchUser();
+  }, [id]);
+
+  const handleInputChange = (field: keyof typeof formData, value: string) => {
+    setFormData((prev) => ({ ...prev, [field]: value }));
+  };
 
   // ------------------- Submit -------------------
   const handleSubmit = async (e: React.FormEvent) => {
@@ -63,17 +87,25 @@ export default function CreateRole() {
     setSaving(true);
 
     try {
-      await api.post("/users", {
-        ...formData,
-        role_id: Number(formData.role_id),
-      });
-      showSuccess("User created successfully!");
-      router.push("/user");
+      const response = await api.post(`/users/${id}?_method=PUT`, formData);
+      showSuccess("User updated successfully!");
+      router.push("/pengguna");
     } catch (error: any) {
       if (error.response?.status === 422) {
-        showValidationErrors(error.response.data.errors);
+        // Handle validation errors
+        if (error.response.data.errors) {
+          showValidationErrors(error.response.data.errors);
+        } else if (error.response.data.error_message) {
+          // Handle error message from repository (e.g., incorrect current password)
+          showError(error.response.data.error_message);
+        } else {
+          showError("Validation failed. Please check your input.");
+        }
+      } else if (error.response?.data?.error_message) {
+        // Handle error message from repository
+        showError(error.response.data.error_message);
       } else {
-        showError("Failed to create user!");
+        showError(error.response?.data?.message || "Failed to update user!");
       }
     } finally {
       setSaving(false);
@@ -81,7 +113,7 @@ export default function CreateRole() {
   };
 
   // ------------------- Render -------------------
-  return (
+   return (
     <ProtectedRoute>
       <Layout>
         <div className="container mx-auto px-6 sm:px-16 py-8">
@@ -90,22 +122,22 @@ export default function CreateRole() {
           <form onSubmit={handleSubmit} className="space-y-8">
             <div className="flex space-x-6">
               <LabelInputContainer>
-                <Label htmlFor="name">Name</Label>
+                <Label htmlFor="name">Nama</Label>
                 <Input
                   id="name"
                   value={formData.name}
                   onChange={(e) => handleInputChange("name", e.target.value)}
-                  placeholder="Enter name"
+                  placeholder="Masukkan nama"
                   required
                 />
               </LabelInputContainer>
               <LabelInputContainer>
-                <Label htmlFor="name">User Name</Label>
+                <Label htmlFor="username">Username</Label>
                 <Input
                   id="username"
                   value={formData.username}
                   onChange={(e) => handleInputChange("username", e.target.value)}
-                  placeholder="Enter user name"
+                  placeholder="Masukkan username"
                   required
                 />
               </LabelInputContainer>
@@ -117,21 +149,51 @@ export default function CreateRole() {
                   id="email"
                   value={formData.email}
                   onChange={(e) => handleInputChange("email", e.target.value)}
-                  placeholder="Enter email"
+                  placeholder="Masukkan email"
                   required
                 />
               </LabelInputContainer>
               <LabelInputContainer>
-                <Label htmlFor="password">Password</Label>
+                <Label htmlFor="current_password">Password Saat Ini</Label>
                 <Input
-                  id="password"
-                  value={formData.password}
-                  onChange={(e) => handleInputChange("password", e.target.value)}
-                  placeholder="Enter password"
-                  required
+                  id="current_password"
+                  type="password"
+                  value={formData.current_password}
+                  onChange={(e) => handleInputChange("current_password", e.target.value)}
+                  placeholder="Masukkan password saat ini untuk mengubah password"
                 />
+                <p className="text-xs text-muted-foreground mt-1">
+                  Kosongkan password fields jika tidak ingin mengubah password
+                </p>
               </LabelInputContainer>
             </div>
+            {(formData.current_password || formData.password || formData.password_confirmation) && (
+              <div className="flex space-x-6">
+                <LabelInputContainer>
+                  <Label htmlFor="password">Password Baru</Label>
+                  <Input
+                    id="password"
+                    type="password"
+                    value={formData.password}
+                    onChange={(e) => handleInputChange("password", e.target.value)}
+                    placeholder="Enter new password"
+                  />
+                  <p className="text-xs text-muted-foreground mt-1">
+                    Harus memiliki minimal 8 karakter dengan kombinasi huruf besar, huruf kecil, angka, dan karakter khusus (@$!%*?&)
+                  </p>
+                </LabelInputContainer>
+                <LabelInputContainer>
+                  <Label htmlFor="password_confirmation">Konfirmasi Password Baru</Label>
+                  <Input
+                    id="password_confirmation"
+                    type="password"
+                    value={formData.password_confirmation}
+                    onChange={(e) => handleInputChange("password_confirmation", e.target.value)}
+                    placeholder="Masukkan konfirmasi password baru"
+                  />
+                </LabelInputContainer>
+              </div>
+            )}
               <LabelInputContainer>
                 <Label htmlFor="role_id">Role</Label>
                 <Select
@@ -139,7 +201,7 @@ export default function CreateRole() {
                   onValueChange={(value) => handleInputChange("role_id", value)}
                 >
                   <SelectTrigger id="role_id" className="w-full">
-                    <SelectValue placeholder="Select Role" />
+                    <SelectValue placeholder="Pilih Role" />
                   </SelectTrigger>
                   <SelectContent>
                     {roleResponse.length > 0 ? (
@@ -160,7 +222,7 @@ export default function CreateRole() {
                 className="cursor-pointer px-6"
                 disabled={saving}
               >
-                {saving ? "Saving..." : "Create User"}
+                {saving ? "Saving..." : "Update User"}
               </Button>
             </div>
           </form>
