+++ frontend/src/app/rekap-keuangan/kas-bank/page.tsx
@@ -250,7 +250,7 @@ export default function FundCashBankPage() {
                 <span className="hidden sm:inline">Refresh</span>
               </button>
               <button
-                onClick={() => router.push("/rekap-keuangan/kas-bank/create")}
+                onClick={() => router.push("/rekap-keuangan/kas-bank/tambah")}
                 className="inline-flex items-center gap-2 rounded-xl bg-neutral-900 text-white px-2 md:px-4 py-2 text-sm font-semibold shadow hover:shadow-md hover:bg-neutral-800 cursor-pointer"
               >
                 <Plus className="h-4 w-4" />
