+++ frontend/src/app/role/[id]/page.tsx
@@ -100,7 +100,7 @@ export default function EditRole() {
     <ProtectedRoute>
       <Layout>
         <div className="container mx-auto px-6 sm:px-16 py-8">
-          <h1 className="text-2xl font-bold mb-6">Edit Role</h1>
+          <h1 className="text-2xl font-bold mb-6">Edit Hak Akses</h1>
 
           {loading ? (
             <p>Loading...</p>
@@ -108,19 +108,19 @@ export default function EditRole() {
             <form onSubmit={handleSubmit} className="space-y-8">
               {/* Role Name */}
               <LabelInputContainer>
-                <Label htmlFor="name">Role Name</Label>
+                <Label htmlFor="name">Nama Hak Akses</Label>
                 <Input
                   id="name"
                   value={formData.name}
                   onChange={(e) => handleInputChange("name", e.target.value)}
-                  placeholder="Enter role name"
+                  placeholder="Masukkan nama hak akses"
                   required
                 />
               </LabelInputContainer>
 
               {/* Permissions */}
               <div>
-                <h2 className="text-lg font-semibold mb-2">Permissions</h2>
+                <h2 className="text-lg font-semibold mb-2">Izin Akses</h2>
                 <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-2 border rounded-lg p-4 max-h-[400px] overflow-y-auto">
                   {allPermissions.map((perm) => (
                     <Label key={perm} className="flex items-center space-x-2">
